+++ MVP Module/Tests/Feed UI/FeedUISnapshotTests.swift
@@ -52,6 +52,6 @@ class FeedUISnapshotTests: XCTestCase {
 
 private extension FeedViewController {
 	func display(errorMessage: String) {
-		errorView?.show(message: errorMessage)
+		refreshController?.errorView?.show(message: errorMessage)
 	}
 }
